Added hero image and text
