+++ app/page.tsx
@@ -2,15 +2,18 @@ import Image from "next/image";
 
 import { getCurrentUser } from "@/lib/auth/checkUser";
 import { redirect } from "next/navigation";
+import HeroDesktop from "@/components/shared/home/hero/HeroDesktop";
+import HeroMobile from "@/components/shared/home/hero/HeroMobile";
 
 export default async function Home() {
-  // const myCurrentUser = await getCurrentUser();
-  // if (myCurrentUser?.onboardingComplete) {
-  //   redirect("/my-stuff");
-  // }
+  const myCurrentUser = await getCurrentUser();
+  if (myCurrentUser?.onboardingComplete) {
+    redirect("/my-stuff");
+  }
   return (
     <section className="flex flex-col items-center justify-between">
-      <h1>hello</h1>
+      <HeroDesktop />
+      <HeroMobile />
     </section>
   );
 }
